Change: Separate fontcache implementations.
