+++ src/fontcache/CMakeLists.txt
@@ -1,0 +1,11 @@
+add_files(
+    freetypefontcache.cpp
+    CONDITION Freetype_FOUND
+)
+
+add_files(
+    spritefontcache.cpp
+    spritefontcache.h
+    truetypefontcache.cpp
+    truetypefontcache.h
+)
